+++ src/test/java/com/epam/reportportal/junit/features/exception/ExpectedExceptionNotThrownTest.java
@@ -24,6 +24,7 @@ public class ExpectedExceptionNotThrownTest {
 
 	public static final String ERROR_MESSAGE = "My error message";
 
+	@SuppressWarnings("deprecation")
 	@Rule
 	public ExpectedException thrown = ExpectedException.none();