+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## [Unreleased]
 ### Changed
 - JUnit-Foundation updated on [12.5.0](https://github.com/sbabcoc/JUnit-Foundation/releases/tag/junit-foundation-12.5.0)
-- Client version updated on [5.0.17](https://github.com/reportportal/client-java/releases/tag/5.0.17)
+- Client version updated on [5.0.18](https://github.com/reportportal/client-java/releases/tag/5.0.18)
+- Parent Item status evaluation was moved into the client
+- "Not Issue" issue moved into the client
 
 ## [5.0.0-RC-1]
 ### Added